Add home, quote and navbar component
