+++ src/App.css
@@ -1,8 +1,95 @@
-.container {
-  background-color: #333131;
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;600&display=swap');
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-family: 'Open Sans', sans-serif;
+}
+
+.nav {
+  background-color: #161B22;
+  color: #fff;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 20px;
+  padding: 0 45px;
+}
+
+.logo-title {
+  font-size: 32px;
+}
+
+span {
+  color: #ea580c;
+  margin-left: 10px;
+}
+
+.nav ul {
+  list-style: none;
+  display: flex;
+  gap: 50px;
+  font-size: 20px;
+}
+
+.nav a {
+  color: inherit; 
+  text-decoration: none;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  padding: 15px;
+}
+
+.nav li:hover {
+  border-bottom: 5px solid #ea580c;
+}
+
+.cover {
+  background-color: #0d1117;
+  color: #fff;
+  height: 100vh;
+}
+
+.cover__inner {
+ padding: 45px 30px 0 60px;
+}
+
+.cover-text {
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 30px;
+  margin: 30px 0;
+}
+
+.cover-text-2 {
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 30px;
+}
+
+.quote__random {
+  display: flex;
+  align-items: center;
   justify-content: center;
-  padding: 10px 0 50px 0;
-  margin: 2rem auto;
+  padding-top: 30vh;
+  font-size: 24px;
+  text-align: center;
+  line-height: 48px;
+}
+
+.quote__author {
+  font-style: italic;
+  color: #ea580c;
+}
+
+.title {
+  padding: 50px 0 0 60px
+}
+
+.container__calculator {
+  background-color: #333131;
+  margin: auto;
   border-radius: 40px;
   box-shadow: 25px 25px 75px #363836, -12px -12px 14px #666 inset;
   width: 450px;
