+++ src/components/pages/Home.jsx
@@ -1,0 +1,22 @@
+/* eslint-disable react/jsx-indent */
+/* eslint-disable max-len */
+/* eslint-disable no-trailing-spaces */
+import React from 'react';
+
+function Home() {
+  return (
+    <div className="cover">
+      <div className="cover__inner">
+        <h1 className="cover__title">Welcome to our Page!</h1>
+        <p className="cover-text">
+            Lorem ipsum dolor sit amet consectetur adipisicing elit. Dignissimos velit praesentium et autem, minima esse, similique aspernatur reiciendis dolorum animi saepe aperiam ipsa veritatis! In, cumque tempora architecto at consequatur placeat optio error adipisci molestias recusandae pariatur dolore quo culpa odit non laboriosam quaerat ad esse corporis perferendis quisquam similique, nobis nam! Facere amet omnis cum impedit laudantium suscipit officiis ratione, totam excepturi possimus a cupiditate fugiat animi non esse consectetur laboriosam veniam saepe, consequuntur quas temporibus? Voluptatum, ipsum minus. Impedit ad magni sed, enim placeat aut iure amet ipsa molestiae, dolore porro quaerat magnam provident eaque? Nulla iste deleniti adipisci temporibus eaque, amet impedit laudantium architecto possimus veritatis reiciendis eum, iusto sequi praesentium similique cum quisquam consectetur laborum maiores. Alias dolorum eligendi iusto iste sequi autem asperiores, hic nisi libero ipsam voluptates ab omnis, vero laboriosam ea tempore saepe. Consequuntur repudiandae ea at iste, iure rerum. Rem fugit cum nesciunt tenetur voluptates molestiae deleniti ut praesentium laudantium. Labore, eveniet? Excepturi aperiam non reprehenderit nemo obcaecati numquam quidem officia ipsam dolorem veniam, nostrum eaque, porro sed. Qui
+        </p>
+        <p className="cover-text-2">
+            Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ab consequatur quae, aut veritatis ipsam omnis? Iste asperiores tempora eum modi quaerat deserunt cupiditate, repudiandae ut doloribus molestias sapiente distinctio veritatis accusantium facilis recusandae debitis odit ipsam animi quod hic minus cum delectus ea sed? Placeat dolorum quae repudiandae tempore totam velit ducimus nihil repellendus id deleniti praesentium exercitationem, vero obcaecati vel aut nesciunt tempora dicta delectus pariatur culpa itaque necessitatibus molestiae? Minima, atque error repellat repellendus quae distinctio quis pariatur soluta iusto! Alias vitae enim possimus exercitationem nemo delectus repudiandae quod vel numquam quaerat? Quidem, magnam. Earum cum incidunt, alias eum assumenda, accusantium illum quae, suscipit eaque iure laborum vero. Numquam repellat quaerat voluptatibus eveniet nihil fugit similique praesentium consectetur, ratione nostrum delectus animi ut harum natus soluta id perferendis iusto ad minima esse reiciendis excepturi voluptatem facere sequi. Animi assumenda libero ea vel magni illo nobis necessitatibus nesciunt? Velit autem quas delectus commodi cum consequatur, voluptatem earum perspiciatis. Laborum odit eveniet porro sit perferendis mollitia consequatur ullam id soluta ipsa dolore animi iste nesciunt aliquid deleniti, beatae amet accusantium aut? Beatae dolores distinctio rerum aspernatur reprehenderit nam harum odio officia accusantium, deleniti placeat vero fuga qui architecto nisi animi! 
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default Home;
