+++ src/components/pages/Quote.jsx
@@ -1,0 +1,22 @@
+/* eslint-disable react/jsx-indent */
+/* eslint-disable max-len */
+/* eslint-disable no-trailing-spaces */
+import React from 'react';
+
+function Quote() {
+  return (
+    <div className="cover">
+      <div className="quote__random">
+        <p>
+            Mathematics is not about numbers, equations, computations, or algorithms: It is about understanding. 
+        <br /> 
+        <span className="quote__author">
+            - William Paul Thurston
+        </span>
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default Quote;
